# backend/app/modules/market_data/api/asset/router.py
"""Asset, asset type, FII/ETF, fixed income, treasury bond, event and exchange routes."""

from fastapi import APIRouter, Depends, Query

from app.composition.market_data import get_asset_service
from app.modules.market_data.api.asset.schemas import (
    AssetCreate,
    AssetDetailsOut,
    AssetEvent,
    AssetType,
    AssetUpdate,
    ExchangeOut,
    FavoriteAsset,
    FixedIncomeAsset,
    FixedIncomeType,
    TreasuryBondTypeOut,
)
from app.modules.market_data.service.asset_service import AssetService
from app.modules.users.domain import User
from app.modules.users.views import current_active_user, current_superuser

router = APIRouter(prefix='/asset', tags=['Asset'])
MAX_FAVORITE_ASSET_FILTER = 1000


# ---------------------------------------------------------------------------
# Auxiliary listings (declared first to avoid clashing with /{asset_id})
# ---------------------------------------------------------------------------
@router.get('/type', response_model=list[AssetType])
async def list_asset_types(
    service: AssetService = Depends(get_asset_service),
):
    """List all asset types."""
    return await service.list_asset_types()


@router.post('/fixed_income')
async def create_fixed_income(
    fixed_income: FixedIncomeAsset,
    service: AssetService = Depends(get_asset_service),
):
    """Create a new fixed income asset."""
    return await service.create_fixed_income(fixed_income.model_dump())


@router.get('/fixed_income/type', response_model=list[FixedIncomeType])
async def list_fixed_income_types(
    service: AssetService = Depends(get_asset_service),
):
    """List all fixed income types."""
    return await service.list_fixed_income_types()


@router.get('/fii/segment')
async def list_fii_segments(
    service: AssetService = Depends(get_asset_service),
):
    """List FII segments."""
    return await service.list_fii_segments()


@router.get('/etf/segment')
async def list_etf_segments(
    service: AssetService = Depends(get_asset_service),
):
    """List all ETF segments."""
    return await service.list_etf_segments()


@router.get('/treasury_bond/type', response_model=list[TreasuryBondTypeOut])
async def list_treasury_bond_types(
    service: AssetService = Depends(get_asset_service),
):
    """List all treasury bond types."""
    return await service.list_treasury_bond_types()


@router.get('/exchange', response_model=list[ExchangeOut])
async def list_exchanges(
    service: AssetService = Depends(get_asset_service),
):
    """List all exchanges."""
    return await service.list_exchanges()


@router.get('/favorites', response_model=list[FavoriteAsset])
async def list_favorite_assets(
    limit: int = Query(default=8, ge=1, le=24),
    asset_type_id: int | None = Query(default=None, ge=1),
    asset_ids: list[int] | None = Query(default=None, max_length=MAX_FAVORITE_ASSET_FILTER),
    user: User = Depends(current_active_user),
    service: AssetService = Depends(get_asset_service),
):
    """The assets this user opens most often."""
    return await service.list_favorite_assets(
        user.id,
        limit,
        asset_type_id,
        asset_ids=asset_ids,
    )


@router.post('/{asset_id}/visit', status_code=204)
async def record_asset_visit(
    asset_id: int,
    user: User = Depends(current_active_user),
    service: AssetService = Depends(get_asset_service),
):
    """Count a visit, which is what ranks the favourites."""
    await service.record_visit(user.id, asset_id)


# ---------------------------------------------------------------------------
# Asset Event sub-resource
# ---------------------------------------------------------------------------
@router.get(
    '/event',
    response_model=list[AssetEvent],
    dependencies=[Depends(current_superuser)],
)
async def list_events(
    service: AssetService = Depends(get_asset_service),
):
    """List all asset events."""
    return await service.list_events()


@router.post('/event', dependencies=[Depends(current_superuser)])
async def create_event(
    event: AssetEvent,
    service: AssetService = Depends(get_asset_service),
):
    """Create a new asset event."""
    return await service.create_event(event)


@router.put('/event/{event_id}', dependencies=[Depends(current_superuser)])
async def update_event(
    event_id: int,
    event: AssetEvent,
    service: AssetService = Depends(get_asset_service),
):
    """Update an asset event."""
    payload = event.model_copy(update={'id': event_id})
    return await service.update_event(payload)


@router.delete('/event/{event_id}', dependencies=[Depends(current_superuser)])
async def delete_event(
    event_id: int,
    service: AssetService = Depends(get_asset_service),
):
    """Delete an asset event."""
    await service.delete_event(event_id)
    return {'message': 'OK'}


# ---------------------------------------------------------------------------
# Asset CRUD
# ---------------------------------------------------------------------------
@router.get('')
async def list_assets(
    service: AssetService = Depends(get_asset_service),
):
    """List all assets (cached for 24h)."""
    return await service.list_assets()


@router.post('')
async def create_asset(
    data: AssetCreate,
    service: AssetService = Depends(get_asset_service),
):
    """Create a new asset with subclass data."""
    return await service.create_asset(data.model_dump())


@router.get('/{asset_id}', response_model=AssetDetailsOut)
async def get_asset(
    asset_id: int,
    service: AssetService = Depends(get_asset_service),
):
    """Get a single asset with all details."""
    return await service.get_asset(asset_id)


@router.put('/{asset_id}')
async def update_asset(
    asset_id: int,
    data: AssetUpdate,
    service: AssetService = Depends(get_asset_service),
):
    """Update an asset with subclass data."""
    return await service.update_asset({**data.model_dump(), 'id': asset_id})


@router.delete('/{asset_id}')
async def delete_asset(
    asset_id: int,
    service: AssetService = Depends(get_asset_service),
):
    """Delete an asset by ID."""
    return await service.delete_asset(asset_id)