# backend/tests/contract/openapi.json
{
 "components": {
  "schemas": {
...
   "AssetClass": {
    "properties": {
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     }
    },
    "required": [
     "id",
     "name"
    ],
    "title": "AssetClass",
    "type": "object"
   },
...
   "AssetCreate": {
    "properties": {
     "anbima_category": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Anbima Category"
     },
     "anbima_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Anbima Code"
     },
     "anbima_code_class": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Anbima Code Class"
     },
     "asset_type_id": {
      "title": "Asset Type Id",
      "type": "integer"
     },
     "country": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Country"
     },
     "etf_segment_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Etf Segment Id"
     },
     "exchange_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Exchange Id"
     },
     "fee": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fee"
     },
     "fii_segment_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fii Segment Id"
     },
     "fixed_income_type_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fixed Income Type Id"
     },
     "index_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Index Id"
     },
     "industry": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Industry"
     },
     "legal_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Legal Id"
     },
     "maturity_date": {
      "anyOf": [
       {
        "format": "date",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Maturity Date"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "sector": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sector"
     },
     "ticker": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ticker"
     },
     "treasury_bond_type_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Treasury Bond Type Id"
     }
    },
    "required": [
     "name",
     "asset_type_id"
    ],
    "title": "AssetCreate",
    "type": "object"
   },
   "AssetDetailsOut": {
    "properties": {
     "asset_type": {
      "$ref": "#/components/schemas/AssetTypeOut"
     },
     "asset_type_id": {
      "title": "Asset Type Id",
      "type": "integer"
     },
     "etf": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ETFOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "exchange_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Exchange Id"
     },
     "fii": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/FIIOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "fixed_income": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/FixedIncomeOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "fund": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/InvestmentFundOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "stock": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/StockOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "ticker": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ticker"
     },
     "treasury_bond": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/TreasuryBondOut"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "required": [
     "id",
     "ticker",
     "name",
     "asset_type_id",
     "asset_type"
    ],
    "title": "AssetDetailsOut",
    "type": "object"
   },
...
   "AssetEvent": {
    "properties": {
     "asset_id": {
      "title": "Asset Id",
      "type": "integer"
     },
     "date": {
      "format": "date",
      "title": "Date",
      "type": "string"
     },
     "factor": {
      "title": "Factor",
      "type": "number"
     },
     "id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Id"
     },
     "type": {
      "title": "Type",
      "type": "string"
     }
    },
    "required": [
     "id",
     "asset_id",
     "date",
     "factor",
     "type"
    ],
    "title": "AssetEvent",
    "type": "object"
   },
...
   "AssetQuoteHistoryResponse": {
    "properties": {
     "asset_type_id": {
      "title": "Asset Type Id",
      "type": "integer"
     },
     "cagr": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/HistoricalCagrResponse"
       },
       {
        "type": "null"
       }
      ]
     },
     "currency": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Currency"
     },
     "logo_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Logo Url"
     },
     "quotes": {
      "items": {
       "$ref": "#/components/schemas/QuoteResponse"
      },
      "title": "Quotes",
      "type": "array"
     },
     "source": {
      "title": "Source",
      "type": "string"
     },
     "ticker": {
      "title": "Ticker",
      "type": "string"
     }
    },
    "required": [
     "ticker",
     "asset_type_id",
     "source",
     "quotes"
    ],
    "title": "AssetQuoteHistoryResponse",
    "type": "object"
   },
...
   "AssetType": {
    "properties": {
     "asset_class": {
      "$ref": "#/components/schemas/AssetClass"
     },
     "asset_class_id": {
      "title": "Asset Class Id",
      "type": "integer"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "short_name": {
      "title": "Short Name",
      "type": "string"
     }
    },
    "required": [
     "id",
     "short_name",
     "name",
     "asset_class_id",
     "asset_class"
    ],
    "title": "AssetType",
    "type": "object"
   },
...
   "ExchangeOut": {
    "properties": {
     "code": {
      "title": "Code",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     }
    },
    "required": [
     "id",
     "code",
     "name"
    ],
    "title": "ExchangeOut",
    "type": "object"
   },
...
   "FIIOut": {
    "properties": {
     "asset_id": {
      "title": "Asset Id",
      "type": "integer"
     },
     "segment": {
      "$ref": "#/components/schemas/FIISegmentOut"
     },
     "segment_id": {
      "title": "Segment Id",
      "type": "integer"
     }
    },
    "required": [
     "asset_id",
     "segment_id",
     "segment"
    ],
    "title": "FIIOut",
    "type": "object"
   },
...
   "FavoriteAsset": {
    "description": "An asset the user returns to, ranked by how often they open it.",
    "properties": {
     "asset_type": {
      "$ref": "#/components/schemas/FavoriteAssetType"
     },
     "asset_type_id": {
      "title": "Asset Type Id",
      "type": "integer"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "last_visited_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Visited At"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "ticker": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ticker"
     },
     "visit_count": {
      "title": "Visit Count",
      "type": "integer"
     }
    },
    "required": [
     "id",
     "ticker",
     "name",
     "asset_type_id",
     "asset_type",
     "visit_count",
     "last_visited_at"
    ],
    "title": "FavoriteAsset",
    "type": "object"
   },
   "FavoriteAssetType": {
    "properties": {
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "short_name": {
      "title": "Short Name",
      "type": "string"
     }
    },
    "required": [
     "id",
     "short_name",
     "name"
    ],
    "title": "FavoriteAssetType",
    "type": "object"
   },
   "FixedIncomeAsset": {
    "properties": {
     "asset_type_id": {
      "title": "Asset Type Id",
      "type": "integer"
     },
     "fee": {
      "title": "Fee",
      "type": "number"
     },
     "fixed_income_type_id": {
      "title": "Fixed Income Type Id",
      "type": "integer"
     },
     "index_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Index Id"
     },
     "maturity_date": {
      "format": "date",
      "title": "Maturity Date",
      "type": "string"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "ticker": {
      "title": "Ticker",
      "type": "string"
     }
    },
    "required": [
     "name",
     "ticker",
     "maturity_date",
     "fee",
     "index_id",
     "fixed_income_type_id",
     "asset_type_id"
    ],
    "title": "FixedIncomeAsset",
    "type": "object"
   },
...
   "FixedIncomeType": {
    "properties": {
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     }
    },
    "required": [
     "id",
     "name",
     "description"
    ],
    "title": "FixedIncomeType",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "TreasuryBondTypeOut": {
    "properties": {
     "code": {
      "title": "Code",
      "type": "string"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     }
    },
    "required": [
     "id",
     "code",
     "name",
     "description"
    ],
    "title": "TreasuryBondTypeOut",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "auth/jwt/login"
     }
    },
    "type": "oauth2"
   }
  }
 },
 "info": {
  "description": "API for portfolio consolidation",
  "title": "My Stonks API",
  "version": "1.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/market_data/asset": {
   "get": {
    "description": "List all assets (cached for 24h).",
    "operationId": "list_assets_market_data_asset_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "List Assets",
    "tags": [
     "Asset"
    ]
   },
   "post": {
    "description": "Create a new asset with subclass data.",
    "operationId": "create_asset_market_data_asset_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AssetCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Create Asset",
    "tags": [
     "Asset"
    ]
   }
  },
  "/market_data/asset/etf/segment": {
   "get": {
    "description": "List all ETF segments.",
    "operationId": "list_etf_segments_market_data_asset_etf_segment_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "List Etf Segments",
    "tags": [
     "Asset"
    ]
   }
  },
  "/market_data/asset/event": {
   "get": {
    "description": "List all asset events.",
    "operationId": "list_events_market_data_asset_event_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/AssetEvent"
         },
         "title": "Response List Events Market Data Asset Event Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "List Events",
    "tags": [
     "Asset"
    ]
   },
   "post": {
    "description": "Create a new asset event.",
    "operationId": "create_event_market_data_asset_event_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AssetEvent"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Create Event",
    "tags": [
     "Asset"
    ]
   }
  },
  "/market_data/asset/event/{event_id}": {
   "delete": {
    "description": "Delete an asset event.",
    "operationId": "delete_event_market_data_asset_event__event_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "event_id",
      "required": true,
      "schema": {
       "title": "Event Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Delete Event",
    "tags": [
     "Asset"
    ]
   },
   "put": {
    "description": "Update an asset event.",
    "operationId": "update_event_market_data_asset_event__event_id__put",
    "parameters": [
     {
      "in": "path",
      "name": "event_id",
      "required": true,
      "schema": {
       "title": "Event Id",
       "type": "integer"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AssetEvent"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Update Event",
    "tags": [
     "Asset"
    ]
   }
  },
  "/market_data/asset/exchange": {
   "get": {
    "description": "List all exchanges.",
    "operationId": "list_exchanges_market_data_asset_exchange_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/ExchangeOut"
         },
         "title": "Response List Exchanges Market Data Asset Exchange Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "List Exchanges",
    "tags": [
     "Asset"
    ]
   }
  },
  "/market_data/asset/favorites": {
   "get": {
    "description": "The assets this user opens most often.",
    "operationId": "list_favorite_assets_market_data_asset_favorites_get",
    "parameters": [
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 8,
       "maximum": 24,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "asset_type_id",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "minimum": 1,
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Asset Type Id"
      }
     },
     {
      "in": "query",
      "name": "asset_ids",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "items": {
          "type": "integer"
         },
         "maxItems": 1000,
         "type": "array"
        },
        {
         "type": "null"
        }
       ],
       "title": "Asset Ids"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/FavoriteAsset"
         },
         "title": "Response List Favorite Assets Market Data Asset Favorites Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "List Favorite Assets",
    "tags": [
     "Asset"
    ]
   }
  },
  "/market_data/asset/fii/segment": {
   "get": {
    "description": "List FII segments.",
    "operationId": "list_fii_segments_market_data_asset_fii_segment_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "List Fii Segments",
    "tags": [
     "Asset"
    ]
   }
  },
  "/market_data/asset/fixed_income": {
   "post": {
    "description": "Create a new fixed income asset.",
    "operationId": "create_fixed_income_market_data_asset_fixed_income_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/FixedIncomeAsset"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Create Fixed Income",
    "tags": [
     "Asset"
    ]
   }
  },
  "/market_data/asset/fixed_income/type": {
   "get": {
    "description": "List all fixed income types.",
    "operationId": "list_fixed_income_types_market_data_asset_fixed_income_type_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/FixedIncomeType"
         },
         "title": "Response List Fixed Income Types Market Data Asset Fixed Income Type Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "List Fixed Income Types",
    "tags": [
     "Asset"
    ]
   }
  },
  "/market_data/asset/treasury_bond/type": {
   "get": {
    "description": "List all treasury bond types.",
    "operationId": "list_treasury_bond_types_market_data_asset_treasury_bond_type_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/TreasuryBondTypeOut"
         },
         "title": "Response List Treasury Bond Types Market Data Asset Treasury Bond Type Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "List Treasury Bond Types",
    "tags": [
     "Asset"
    ]
   }
  },
  "/market_data/asset/type": {
   "get": {
    "description": "List all asset types.",
    "operationId": "list_asset_types_market_data_asset_type_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/AssetType"
         },
         "title": "Response List Asset Types Market Data Asset Type Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "List Asset Types",
    "tags": [
     "Asset"
    ]
   }
  },
  "/market_data/asset/{asset_id}": {
   "delete": {
    "description": "Delete an asset by ID.",
    "operationId": "delete_asset_market_data_asset__asset_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "asset_id",
      "required": true,
      "schema": {
       "title": "Asset Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Delete Asset",
    "tags": [
     "Asset"
    ]
   },
   "get": {
    "description": "Get a single asset with all details.",
    "operationId": "get_asset_market_data_asset__asset_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "asset_id",
      "required": true,
      "schema": {
       "title": "Asset Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AssetDetailsOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Asset",
    "tags": [
     "Asset"
    ]
   },
   "put": {
    "description": "Update an asset with subclass data.",
    "operationId": "update_asset_market_data_asset__asset_id__put",
    "parameters": [
     {
      "in": "path",
      "name": "asset_id",
      "required": true,
      "schema": {
       "title": "Asset Id",
       "type": "integer"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AssetUpdate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Update Asset",
    "tags": [
     "Asset"
    ]
   }
  },
  "/market_data/asset/{asset_id}/visit": {
   "post": {
    "description": "Count a visit, which is what ranks the favourites.",
    "operationId": "record_asset_visit_market_data_asset__asset_id__visit_post",
    "parameters": [
     {
      "in": "path",
      "name": "asset_id",
      "required": true,
      "schema": {
       "title": "Asset Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Record Asset Visit",
    "tags": [
     "Asset"
    ]
   }
  },
...
  "/market_data/quotes/asset/{asset_id}": {
   "get": {
    "description": "Quote history for one asset, from storage when it exists.\n\nPrices come back in ``currency``, converted through the USD/BRL history when\nthe asset is not quoted in it.",
    "operationId": "get_asset_quote_history_market_data_quotes_asset__asset_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "asset_id",
      "required": true,
      "schema": {
       "title": "Asset Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "start_date",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "format": "date",
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Start Date"
      }
     },
     {
      "in": "query",
      "name": "currency",
      "required": false,
      "schema": {
       "default": "BRL",
       "title": "Currency",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AssetQuoteHistoryResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Asset Quote History",
    "tags": [
     "Quotes"
    ]
   }
  },
...
 }
}